pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "connect4-alphazero"
version = "0.1.0"
description = "AlphaZero implementation for Connect4 and TicTacToe"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Robi Bhattacharjee" }
]

dependencies = [
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "pygame>=2.5.0",
    "jaxtyping>=0.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
